RTABmap depth_only launch
TF trees kept breaking and data was not shared with rtab nodes because it subscribes to rgb inputs which dont exist. Also initially a false odom is required and fixed frame of robot, not camera, is required which is added by this file. This file doesn't launch Synexens CS20 sensor, it has to be launched separately because it was built in a different workspace of ros_ws and hence won't respoond to a package in rtabmap_ws.